Rename ClassObject to ParseObject
