+++ src/GraphQL/loaders/defaultGraphQLMutations.js
@@ -1,7 +1,7 @@
 import * as filesMutations from './filesMutations';
 import * as usersMutations from './usersMutations';
 import * as functionsMutations from './functionsMutations';
-import * as classSchemaMutations from './classSchemaMutations';
+import * as classSchemaMutations from './schemaMutations';
 
 const load = parseGraphQLSchema => {
   filesMutations.load(parseGraphQLSchema);
